+++ cpp11test/R/cpp11.R
@@ -20,6 +20,26 @@ my_message <- function(mystring, myarg) {
   invisible(.Call(`_cpp11test_my_message`, mystring, myarg))
 }
 
+my_stop_n1fmt <- function(mystring) {
+  invisible(.Call(`_cpp11test_my_stop_n1fmt`, mystring))
+}
+
+my_stop_n2fmt <- function(mystring, myarg) {
+  invisible(.Call(`_cpp11test_my_stop_n2fmt`, mystring, myarg))
+}
+
+my_warning_n1fmt <- function(mystring) {
+  invisible(.Call(`_cpp11test_my_warning_n1fmt`, mystring))
+}
+
+my_warning_n2fmt <- function(mystring, myarg) {
+  invisible(.Call(`_cpp11test_my_warning_n2fmt`, mystring, myarg))
+}
+
+my_message_nfmt <- function(mystring, myarg) {
+  invisible(.Call(`_cpp11test_my_message_nfmt`, mystring, myarg))
+}
+
 remove_altrep <- function(x) {
   .Call(`_cpp11test_remove_altrep`, x)
 }