+++ cpp11test/src/errors.cpp
@@ -2,12 +2,15 @@
 #include "cpp11/function.hpp"
 using namespace cpp11;
 
-[[cpp11::register]] void my_stop(std::string mystring, int myarg) {
-  cpp11::stop(mystring, myarg);
+[[cpp11::register]] void my_stop_n1fmt(std::string mystring) { cpp11::stop(mystring); }
+[[cpp11::register]] void my_stop_n2fmt(std::string mystring, std::string myarg) {
+  cpp11::stop(mystring, myarg.c_str());
 }
-
-[[cpp11::register]] void my_warning(std::string mystring, std::string myarg) {
-  cpp11::warning(mystring, myarg);
+[[cpp11::register]] void my_warning_n1fmt(std::string mystring) {
+  cpp11::warning(mystring);
+}
+[[cpp11::register]] void my_warning_n2fmt(std::string mystring, std::string myarg) {
+  cpp11::warning(mystring, myarg.c_str());
 }
 
 [[cpp11::register]] void my_message(std::string mystring, std::string myarg) {