+++ cpp11test/tests/testthat/test_formatted_errors.R
@@ -21,3 +21,20 @@ test_that("cpp11::message formatting works", {
   test2 <- c("great", "super")
   expect_message(my_message("You're {}", test2[2]), "You're super", fixed = TRUE)
 })
+test_that("cpp11::stop works without including the fmt library", {
+  test1 <- "error"
+  expect_error(my_stop_n1fmt("This is a stop"), "This is a stop", fixed = TRUE)
+  expect_error(my_stop_n2fmt("This is an %s", test1), "This is an error", fixed = TRUE)
+})
+test_that("cpp11::warning works without including the fmt library", {
+  test1 <- "warning"
+  expect_warning(my_warning_n1fmt("This is a warning"), "This is a warning", fixed = TRUE)
+  expect_warning(my_warning_n2fmt("This is a %s", test1), "This is a warning", fixed = TRUE)
+})
+test_that("cpp11::message works without including the fmt library", {
+  test1 <- "message"
+  expect_output(my_message_nfmt("This is a %s", test1), "This is a message", fixed = TRUE)
+
+  test2 <- c("great", "super")
+  expect_output(my_message_nfmt("You're %s", test2[2]), "You're super", fixed = TRUE)
+})